Extract Hive bucketing hashing
