+++ presto-hive/src/main/java/io/prestosql/plugin/hive/util/HiveBucketing.java
@@ -18,9 +18,6 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableSet;
-import com.google.common.primitives.Shorts;
-import com.google.common.primitives.SignedBytes;
-import io.airlift.slice.Slice;
 import io.prestosql.plugin.hive.HiveBucketHandle;
 import io.prestosql.plugin.hive.HiveBucketProperty;
 import io.prestosql.plugin.hive.HiveColumnHandle;
@@ -30,16 +27,11 @@ import io.prestosql.plugin.hive.metastore.Table;
 import io.prestosql.spi.Page;
 import io.prestosql.spi.PrestoException;
 import io.prestosql.spi.StandardErrorCode;
-import io.prestosql.spi.block.Block;
 import io.prestosql.spi.connector.ColumnHandle;
 import io.prestosql.spi.predicate.Domain;
 import io.prestosql.spi.predicate.NullableValue;
 import io.prestosql.spi.predicate.TupleDomain;
 import io.prestosql.spi.predicate.ValueSet;
-import io.prestosql.spi.type.Type;
-import org.apache.hadoop.hive.serde2.typeinfo.ListTypeInfo;
-import org.apache.hadoop.hive.serde2.typeinfo.MapTypeInfo;
-import org.apache.hadoop.hive.serde2.typeinfo.PrimitiveTypeInfo;
 import org.apache.hadoop.hive.serde2.typeinfo.TypeInfo;
 
 import java.util.HashMap;
@@ -57,16 +49,10 @@ import static io.prestosql.plugin.hive.HiveErrorCode.HIVE_INVALID_METADATA;
 import static io.prestosql.plugin.hive.util.HiveBucketing.BucketingVersion.BUCKETING_V1;
 import static io.prestosql.plugin.hive.util.HiveBucketing.BucketingVersion.BUCKETING_V2;
 import static io.prestosql.plugin.hive.util.HiveUtil.getRegularColumnHandles;
-import static java.lang.Double.doubleToLongBits;
-import static java.lang.Float.floatToIntBits;
-import static java.lang.Float.intBitsToFloat;
-import static java.lang.Math.toIntExact;
 import static java.lang.String.format;
 import static java.util.Map.Entry;
-import static java.util.Objects.requireNonNull;
 import static java.util.function.Function.identity;
 import static org.apache.hadoop.hive.metastore.api.hive_metastoreConstants.TABLE_BUCKETING_VERSION;
-import static org.apache.hadoop.hive.serde2.objectinspector.PrimitiveObjectInspector.PrimitiveCategory;
 
 public final class HiveBucketing
 {
@@ -119,172 +105,14 @@ public final class HiveBucketing
     static int getBucketHashCode(BucketingVersion bucketingVersion, List<TypeInfo> types, Page page, int position)
     {
         checkArgument(bucketingVersion == BUCKETING_V1);
-        checkArgument(types.size() == page.getChannelCount());
-        int result = 0;
-        for (int i = 0; i < page.getChannelCount(); i++) {
-            int fieldHash = hash(types.get(i), page.getBlock(i), position);
-            result = result * 31 + fieldHash;
-        }
-        return result;
+        return HiveBucketingV1.getBucketHashCode(types, page, position);
     }
 
     @VisibleForTesting
     static int getBucketHashCode(BucketingVersion bucketingVersion, List<TypeInfo> types, Object[] values)
     {
         checkArgument(bucketingVersion == BUCKETING_V1);
-        checkArgument(types.size() == values.length);
-        int result = 0;
-        for (int i = 0; i < values.length; i++) {
-            int fieldHash = hash(types.get(i), values[i]);
-            result = result * 31 + fieldHash;
-        }
-        return result;
-    }
-
-    private static int hash(TypeInfo type, Block block, int position)
-    {
-        // This function mirrors the behavior of function hashCode in
-        // HIVE-12025 ba83fd7bff serde/src/java/org/apache/hadoop/hive/serde2/objectinspector/ObjectInspectorUtils.java
-        // https://github.com/apache/hive/blob/ba83fd7bff/serde/src/java/org/apache/hadoop/hive/serde2/objectinspector/ObjectInspectorUtils.java
-
-        // HIVE-7148 proposed change to bucketing hash algorithms. If that gets implemented, this function will need to change significantly.
-
-        if (block.isNull(position)) {
-            return 0;
-        }
-
-        switch (type.getCategory()) {
-            case PRIMITIVE:
-                PrimitiveTypeInfo typeInfo = (PrimitiveTypeInfo) type;
-                PrimitiveCategory primitiveCategory = typeInfo.getPrimitiveCategory();
-                Type prestoType = requireNonNull(HiveType.getPrimitiveType(typeInfo));
-                switch (primitiveCategory) {
-                    case BOOLEAN:
-                        return prestoType.getBoolean(block, position) ? 1 : 0;
-                    case BYTE:
-                        return SignedBytes.checkedCast(prestoType.getLong(block, position));
-                    case SHORT:
-                        return Shorts.checkedCast(prestoType.getLong(block, position));
-                    case INT:
-                        return toIntExact(prestoType.getLong(block, position));
-                    case LONG:
-                        long bigintValue = prestoType.getLong(block, position);
-                        return (int) ((bigintValue >>> 32) ^ bigintValue);
-                    case FLOAT:
-                        // convert to canonical NaN if necessary
-                        return floatToIntBits(intBitsToFloat(toIntExact(prestoType.getLong(block, position))));
-                    case DOUBLE:
-                        long doubleValue = doubleToLongBits(prestoType.getDouble(block, position));
-                        return (int) ((doubleValue >>> 32) ^ doubleValue);
-                    case STRING:
-                        return hashBytes(0, prestoType.getSlice(block, position));
-                    case VARCHAR:
-                        return hashBytes(1, prestoType.getSlice(block, position));
-                    case DATE:
-                        // day offset from 1970-01-01
-                        return toIntExact(prestoType.getLong(block, position));
-                    case TIMESTAMP:
-                        return hashTimestamp(prestoType.getLong(block, position));
-                    default:
-                        throw new UnsupportedOperationException("Computation of Hive bucket hashCode is not supported for Hive primitive category: " + primitiveCategory);
-                }
-            case LIST:
-                return hashOfList((ListTypeInfo) type, block.getObject(position, Block.class));
-            case MAP:
-                return hashOfMap((MapTypeInfo) type, block.getObject(position, Block.class));
-            default:
-                // TODO: support more types, e.g. ROW
-                throw new UnsupportedOperationException("Computation of Hive bucket hashCode is not supported for Hive category: " + type.getCategory());
-        }
-    }
-
-    private static int hash(TypeInfo type, Object value)
-    {
-        if (value == null) {
-            return 0;
-        }
-
-        switch (type.getCategory()) {
-            case PRIMITIVE:
-                PrimitiveTypeInfo typeInfo = (PrimitiveTypeInfo) type;
-                PrimitiveCategory primitiveCategory = typeInfo.getPrimitiveCategory();
-                switch (primitiveCategory) {
-                    case BOOLEAN:
-                        return (boolean) value ? 1 : 0;
-                    case BYTE:
-                        return SignedBytes.checkedCast((long) value);
-                    case SHORT:
-                        return Shorts.checkedCast((long) value);
-                    case INT:
-                        return toIntExact((long) value);
-                    case LONG:
-                        long bigintValue = (long) value;
-                        return (int) ((bigintValue >>> 32) ^ bigintValue);
-                    case FLOAT:
-                        // convert to canonical NaN if necessary
-                        return floatToIntBits(intBitsToFloat(toIntExact((long) value)));
-                    case DOUBLE:
-                        long doubleValue = doubleToLongBits((double) value);
-                        return (int) ((doubleValue >>> 32) ^ doubleValue);
-                    case STRING:
-                        return hashBytes(0, (Slice) value);
-                    case VARCHAR:
-                        return hashBytes(1, (Slice) value);
-                    case DATE:
-                        // day offset from 1970-01-01
-                        return toIntExact((long) value);
-                    case TIMESTAMP:
-                        return hashTimestamp((long) value);
-                    default:
-                        throw new UnsupportedOperationException("Computation of Hive bucket hashCode is not supported for Hive primitive category: " + primitiveCategory);
-                }
-            case LIST:
-                return hashOfList((ListTypeInfo) type, (Block) value);
-            case MAP:
-                return hashOfMap((MapTypeInfo) type, (Block) value);
-            default:
-                // TODO: support more types, e.g. ROW
-                throw new UnsupportedOperationException("Computation of Hive bucket hashCode is not supported for Hive category: " + type.getCategory());
-        }
-    }
-
-    @SuppressWarnings("NumericCastThatLosesPrecision")
-    private static int hashTimestamp(long epochMillis)
-    {
-        long seconds = (Math.floorDiv(epochMillis, 1000L) << 30);
-        long nanos = Math.floorMod(epochMillis, 1000) * 1_000_000L;
-        long secondsAndNanos = seconds | nanos;
-        return (int) ((secondsAndNanos >>> 32) ^ secondsAndNanos);
-    }
-
-    private static int hashOfMap(MapTypeInfo type, Block singleMapBlock)
-    {
-        TypeInfo keyTypeInfo = type.getMapKeyTypeInfo();
-        TypeInfo valueTypeInfo = type.getMapValueTypeInfo();
-        int result = 0;
-        for (int i = 0; i < singleMapBlock.getPositionCount(); i += 2) {
-            result += hash(keyTypeInfo, singleMapBlock, i) ^ hash(valueTypeInfo, singleMapBlock, i + 1);
-        }
-        return result;
-    }
-
-    private static int hashOfList(ListTypeInfo type, Block singleListBlock)
-    {
-        TypeInfo elementTypeInfo = type.getListElementTypeInfo();
-        int result = 0;
-        for (int i = 0; i < singleListBlock.getPositionCount(); i++) {
-            result = result * 31 + hash(elementTypeInfo, singleListBlock, i);
-        }
-        return result;
-    }
-
-    private static int hashBytes(int initialValue, Slice bytes)
-    {
-        int result = initialValue;
-        for (int i = 0; i < bytes.length(); i++) {
-            result = result * 31 + bytes.getByte(i);
-        }
-        return result;
+        return HiveBucketingV1.getBucketHashCode(types, values);
     }
 
     public static Optional<HiveBucketHandle> getHiveBucketHandle(Table table)
